PR cleanups
