+++ src/pages/Admin/ToolsPage/ToolsPage.scss
@@ -8,15 +8,3 @@
     margin-bottom: $spacing--md;
   }
 }
-
-.Tool {
-  &__input {
-    margin: $spacing--md 0;
-  }
-
-  &__input-label {
-    display: block;
-    opacity: 0.7;
-    margin-bottom: $spacing--sm;
-  }
-}
