+++ src/pages/Admin/ToolsPage/components/Tool/Tool.scss
@@ -1,0 +1,11 @@
+.Tool {
+  &__input {
+    margin: $spacing--md 0;
+  }
+
+  &__input-label {
+    display: block;
+    opacity: 0.7;
+    margin-bottom: $spacing--sm;
+  }
+}
