+++ src/pages/Admin/ToolsPage/components/Tool/index.ts
@@ -1,0 +1,1 @@
+export { Tool } from "./Tool";
